added access control

+++ lsapp/app/Http/Controllers/PostsController.php
@@ -8,6 +8,15 @@ use Illuminate\Http\Request;
 
 class PostsController extends Controller
 {
+      /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' =>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -75,7 +84,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        
         $post= Post::find($id);
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -93,6 +106,7 @@ class PostsController extends Controller
             'body'=>'required'
         ]);
         // edit post 
+        
         $post= Post::find($id);
         $post->title = $request->input('title');
         $post->body=$request->input('body');
@@ -110,6 +124,10 @@ class PostsController extends Controller
     {
             // Delete
         $post= Post::find($id);
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
